backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
colorama==0.4.6
fastapi==0.121.2
h11==0.16.0
idna==3.11
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
sqlmodel==0.0.22

# Database: PostgreSQL Driver for Supabase
psycopg2-binary==2.9.10

# Security: JWT Handling for Clerk
python-jose[cryptography]==3.3.0
httpx==0.26.0